A lot of small tweaks. GUI design nearly finished.

+++ src/main/java/co/uk/silvania/rpgcore/skills/GlobalLevel.java
@@ -76,7 +76,7 @@ public class GlobalLevel extends SkillLevelBase implements IExtendedEntityProper
 		return xpForLevel;
 	}
 	
-	public int xpToNextLevel() {		
+	public float xpToNextLevel() {		
 		return (getXpForLevel(getLevel())) - (int)getXPGlobal();
 	}
